requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
asyncpg
sqlalchemy[asyncio]
passlib[bcrypt]
python-jose
fastapi-users
authlib
stripe
python-jose
python-dotenv
